Moves Items to types
